pyproject.toml
[tool.poetry]
name = "cmr_python_simple"
version = "0.1.0"
description = "Your project description"
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.0"
SQLAlchemy = "2.0.0"
uvicorn = "0.22.0"
python-json-logger = ">=2.0.7,<2.1"
boto3 = ">=1.26.0,<1.27"
redis = ">=4.5.0"
python-dotenv = "^1.0.1"
setuptools = "^75.8.2"

[tool.poetry.group.dev.dependencies]
flake8 = ">=3.9.2"
autoflake = "^2.3.1"
isort = "*"
pytest = "7.3.0"
black = ">=23.3.0"
pylint = "^3.3.4"
pre-commit = "^4.1.0"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''